Refactored my previous commit.
Added the alternate download functionality to the list nzb button and push to sabnzbd button.
Moved the logic into a seporate class and changed the previous made functionality so it wont return empty strings if it fails to find the alternate. If there is no alternate we will serve the original url.