Do not instantiate jQuery objects for tiptip anymore
By using an uglier way - adding raw HTML - we do not trigger the loading of the images up front, and thus prevent paying the preformance penalty of loading the page.

We also increase the default timeout of an image from 400 ms to 800 ms, this feels less responsive, but has the added benefit of not triggering loads of imageloads when just hovering over the list of spots.
Conflicts:
	templates/we1rdo/js/scripts.js